+++ inject/inject-app/src/test/java/com/twitter/inject/app/EmbeddedAppIntegrationJavaTest.java
@@ -1,7 +1,11 @@
 package com.twitter.inject.app;
 
+import com.google.common.collect.ImmutableMap;
+
 import org.junit.Assert;
 import org.junit.Test;
+
+import com.twitter.app.FlagParseException;
 
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -18,6 +22,24 @@ public class EmbeddedAppIntegrationJavaTest extends Assert {
     app.main();
 
     assertEquals(sampleJavaApp.getSampleServiceResponse(), "hi yo");
+  }
+
+  // -help arg exits with non-zero status/throws
+  @Test(expected = Exception.class)
+  public void testEmbeddedHelpArg() {
+    SampleJavaApp sampleJavaApp = new SampleJavaApp();
+
+    EmbeddedApp app = new EmbeddedApp(sampleJavaApp);
+    app.main(ImmutableMap.of(), "-help");
+  }
+
+  // we are just testing that flags get passed, these will all be undefined and should throw
+  @Test(expected = FlagParseException.class)
+  public void testEmbeddedAppFailOnUndefinedFlags() {
+    SampleJavaApp sampleJavaApp = new SampleJavaApp();
+
+    EmbeddedApp app = new EmbeddedApp(sampleJavaApp);
+    app.main(ImmutableMap.of("-flag1", "value1", "-flag2", "value2"));
   }
 
   @Test